@@ -2,7 +2,7 @@ import styles from './App.module.scss';
 import React from "react";
 import Home from "./views/Home/Home";
 import {Header} from "./components/Header/Header";
-import {BrowserRouter, Route, Switch} from "react-router-dom";
+import {BrowserRouter, Route, Routes} from "react-router-dom";
 import {ROUTES} from "./App.model";
 import {Add} from "./views/Add/Add";
 import {Pick} from "./views/Pick/Pick";
@@ -17,23 +17,13 @@ export function App() {
             <BrowserRouter>
                 <Header/>
                 <div className={styles.AppContent}>
-                    <Switch>
-                        <Route exact path={ROUTES.home}>
-                            <Home/>
-                        </Route>
-                        <Route path={ROUTES.add}>
-                            <Add/>
-                        </Route>
-                        <Route path={ROUTES.pick}>
-                            <Pick/>
-                        </Route>
-                        <Route path={ROUTES.play}>
-                            <Play/>
-                        </Route>
-                        <Route path={ROUTES.about}>
-                            <About/>
-                        </Route>
-                    </Switch>
+                    <Routes>
+                        <Route path={ROUTES.home} element={<Home/>}/>
+                        <Route path={ROUTES.add} element={<Add/>}/>
+                        <Route path={ROUTES.pick} element={<Pick/>}/>
+                        <Route path={ROUTES.play} element={<Play/>}/>
+                        <Route path={ROUTES.about} element={<About/>}/>
+                    </Routes>
                 </div>
             </BrowserRouter>
         </div>
