@@ -1,7 +1,7 @@
 import React from 'react';
 import {render, screen} from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import Button from './Button';
+import {Button} from './Button';
 
 describe('<Button />', () => {
     test('it should mount', () => {
