@@ -1,7 +1,7 @@
 import React from 'react';
 import {render, screen} from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import RemoveButton from './RemoveButton';
+import {RemoveButton} from './RemoveButton';
 
 describe('<RemoveButton />', () => {
     test('it should mount', () => {
