@@ -1,7 +1,7 @@
 import React from 'react';
 import {render, screen} from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import TextInput from './TextInput';
+import {TextInput} from './TextInput';
 
 describe('<TextInput />', () => {
     test('it should mount', () => {
