@@ -1,7 +1,7 @@
 import React from 'react';
 import {render, screen} from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import Home from './Home';
+import {Home} from './Home';
 
 describe('<Home />', () => {
     test('it should mount', () => {
