@@ -1,7 +1,7 @@
 import React from 'react';
 import {render, screen} from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import Pick from './Pick';
+import {Pick} from './Pick';
 
 describe('<Pick />', () => {
     test('it should mount', () => {
